Update import/require' pathes for reorganized server side's files
