+++ server/connect.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-const interactiveClient = require('./lib/interactiveClient');
+const interactiveClient = require('./lib/interactive/Client');
 
 interactiveClient();
